+++ api/settings.py
@@ -659,7 +659,7 @@ def setup_django_constants():
             'rest_framework_simplejwt.authentication.JWTAuthentication',
         ),
         'DEFAULT_METADATA_CLASS': 'rest_framework.metadata.SimpleMetadata',
-        'DEFAULT_SCHEMA_CLASS': 'drf_spectacular.openapi.AutoSchema',
+        'DEFAULT_SCHEMA_CLASS': 'api.view.schema.AppAutoSchema.AppAutoSchema',
         'DEFAULT_VERSIONING_CLASS': 'rest_framework.versioning.URLPathVersioning',
         'DEFAULT_VERSION': 'v1',
         'ALLOWED_VERSIONS': 'v1',